requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
jinja2==3.1.*
google-cloud-bigquery>=3.25.0
google-generativeai>=0.8.0
